Test Case: Handling unknown amount of numbers
Unlike previous commit where the test case was to add two numbers separated by comma, in this commit the previous code is refactored to handle multiple numbers in a string which is separated by comma. Arrays.stream method is used to pass sequential input as a parameter. split() method is used to separate the sting into array of string and then it is converted to INT type to perform addition of it.